pyhton-bot-api/requirements.txt
discord.py>=2.3.2   # or py-cord; discord.py 2.x works with Intents
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
